cart-service/requirements.txt
alembic==1.15.2
bcrypt==4.3.0
fastapi==0.115.12
uvicorn==0.34.2
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic-settings==2.9.1
email_validator==2.2.0
fastapi-pagination==0.13.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pytest==8.3.5